pyproject.toml
[tool.poetry]
name = "app-pdf"
version = "0.1.0"
description = ""
authors = ["hotratx <mgviero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyPDF2 = "^2.9.0"
tabula-py = "^2.4.0"
bokeh = "^2.4.3"
plotly = "^5.9.0"
SQLAlchemy = "^1.4.39"
pydantic = "^1.9.1"
dash = "^2.6.1"
dash-bootstrap-components = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
